Move code to LegacyToDomainConverter
Move conversion code from DoctrineDonationRepository to
LegacyToDomainConverter.


Co-authored by: Co <corinna.hillebrand@wikimedia.de>
